# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_hawkbit.ActionStatus:
    properties:
      id:
        type: integer
      status:
        description: '"finished", "running", "error", "canceled"'
        type: string
      type:
        description: '"update", "cancel"'
        type: string
    type: object
  internal_hawkbit.Rollout:
    properties:
      amountGroups:
        type: integer
      distributionSetId:
        type: integer
      id:
        type: integer
      name:
        type: string
      status:
        description: '"creating", "ready", "running", "paused", "finished"'
        type: string
      targetFilterQuery:
        type: string
      totalTargets:
        type: integer
    type: object
  internal_hawkbit.Target:
    properties:
      controllerId:
        type: string
      description:
        type: string
      ipAddress:
        type: string
      lastControllerRequestAt:
        type: string
      name:
        type: string
      updateStatus:
        description: '"registered", "in_sync", "pending", "error"'
        type: string
    type: object
...
host: localhost:6070
info:
  contact: {}
  description: Multi-tenant SaaS platform for satellite device management. Ingests
    MO messages from Iridium/Astrocast, manages devices, SOS escalation, dead man's
    switch, and E2E encryption.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: MeshSat Hub API
  version: "1.1"
paths:
...
  /api/ota/rollouts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Rollout details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_hawkbit.Rollout'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_hawkbit.Rollout'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create firmware rollout
      tags:
      - ota
  /api/ota/rollouts/{id}:
    get:
      parameters:
      - description: Rollout ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_hawkbit.Rollout'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get rollout details
      tags:
      - ota
  /api/ota/rollouts/{id}/pause:
    post:
      parameters:
      - description: Rollout ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Pause a rollout
      tags:
      - ota
  /api/ota/rollouts/{id}/start:
    post:
      parameters:
      - description: Rollout ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start a rollout
      tags:
      - ota
  /api/ota/targets:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_hawkbit.Target'
              type: array
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List OTA targets
      tags:
      - ota
    post:
      consumes:
      - application/json
      parameters:
      - description: Target details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_hawkbit.Target'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_hawkbit.Target'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create OTA target
      tags:
      - ota
  /api/ota/targets/{controllerId}:
    delete:
      parameters:
      - description: Target controller ID
        in: path
        name: controllerId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete OTA target
      tags:
      - ota
    get:
      parameters:
      - description: Target controller ID
        in: path
        name: controllerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_hawkbit.Target'
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get OTA target
      tags:
      - ota
  /api/ota/targets/{controllerId}/actions:
    get:
      parameters:
      - description: Target controller ID
        in: path
        name: controllerId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_hawkbit.ActionStatus'
              type: array
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get target deployment actions
      tags:
      - ota
  /api/ota/targets/{controllerId}/actions/{actionId}:
    delete:
      parameters:
      - description: Target controller ID
        in: path
        name: controllerId
        required: true
        type: string
      - description: Action ID
        in: path
        name: actionId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Cancel deployment action
      tags:
      - ota
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"